+++ spec/spec_helper.rb
@@ -1,15 +1,30 @@
-require "bundler/setup"
-require "capture_migration_sql"
+# frozen_string_literal: true
 
+ENV["BUNDLE_GEMFILE"] ||= File.expand_path("../Gemfile", __dir__)
+
+require "bundler/setup" if File.exist?(ENV["BUNDLE_GEMFILE"])
+
+begin
+  require "simplecov"
+  SimpleCov.start do
+    add_filter ["/spec/"]
+  end
+rescue LoadError
+end
+
+require "logger" # needed for ActiveRecord 6.x and 7.0
 require "active_record"
 
-RSpec.configure do |config|
-  # Enable flags like --only-failures and --next-failure
-  config.example_status_persistence_file_path = ".rspec_status"
+Bundler.require(:default, :test)
 
-  config.expect_with :rspec do |c|
-    c.syntax = :expect
-  end
+require_relative "../lib/capture_migration_sql"
+
+RSpec.configure do |config|
+  config.warnings = true
+  config.disable_monkey_patching!
+  config.default_formatter = "doc" if config.files_to_run.one?
+  config.order = :random
+  Kernel.srand config.seed
 end
 
 ActiveRecord::Base.establish_connection("adapter" => "sqlite3", "database" => ":memory:")